requirements.txt
streamlit>=1.28.0
azure-ai-projects==1.0.0b10
azure-identity>=1.12.0 